+++ kogito-drools/drools-core/src/main/java/org/drools/core/base/evaluators/DuringEvaluatorDefinition.java
@@ -178,7 +178,7 @@ public class DuringEvaluatorDefinition
         String key = isNegated + ":" + parameterText;
         DuringEvaluator eval = this.cache.get( key );
         if ( eval == null ) {
-            Long[] params = TimeIntervalParser.parse( parameterText );
+            long[] params = TimeIntervalParser.parse( parameterText );
             eval = new DuringEvaluator( type,
                                         isNegated,
                                         params,
@@ -238,7 +238,7 @@ public class DuringEvaluatorDefinition
 
         public DuringEvaluator(final ValueType type,
                                final boolean isNegated,
-                               final Long[] parameters,
+                               final long[] parameters,
                                final String paramText) {
             super( type,
                    isNegated ? NOT_DURING : DURING );
@@ -363,7 +363,7 @@ public class DuringEvaluatorDefinition
          *
          * @param parameters
          */
-        private void setParameters(Long[] parameters) {
+        private void setParameters(long[] parameters) {
             if ( parameters == null || parameters.length == 0 ) {
                 // open bounded range
                 this.startMinDev = 1;
@@ -373,21 +373,21 @@ public class DuringEvaluatorDefinition
             } else if ( parameters.length == 1 ) {
                 // open bounded ranges
                 this.startMinDev = 1;
-                this.startMaxDev = parameters[0].longValue();
+                this.startMaxDev = parameters[0];
                 this.endMinDev = 1;
-                this.endMaxDev = parameters[0].longValue();
+                this.endMaxDev = parameters[0];
             } else if ( parameters.length == 2 ) {
                 // open bounded ranges
-                this.startMinDev = parameters[0].longValue();
-                this.startMaxDev = parameters[1].longValue();
-                this.endMinDev = parameters[0].longValue();
-                this.endMaxDev = parameters[1].longValue();
+                this.startMinDev = parameters[0];
+                this.startMaxDev = parameters[1];
+                this.endMinDev = parameters[0];
+                this.endMaxDev = parameters[1];
             } else if ( parameters.length == 4 ) {
                 // open bounded ranges
-                this.startMinDev = parameters[0].longValue();
-                this.startMaxDev = parameters[1].longValue();
-                this.endMinDev = parameters[2].longValue();
-                this.endMaxDev = parameters[3].longValue();
+                this.startMinDev = parameters[0];
+                this.startMaxDev = parameters[1];
+                this.endMinDev = parameters[2];
+                this.endMaxDev = parameters[3];
             } else {
                 throw new RuntimeException( "[During Evaluator]: Not possible to use " + parameters.length + " parameters: '" + paramText + "'" );
             }
