+++ kogito-drools/drools-core/src/main/java/org/drools/core/base/evaluators/MeetsEvaluatorDefinition.java
@@ -155,7 +155,7 @@ public class MeetsEvaluatorDefinition
         String key = isNegated + ":" + parameterText;
         MeetsEvaluator eval = this.cache.get( key );
         if ( eval == null ) {
-            Long[] params = TimeIntervalParser.parse( parameterText );
+            long[] params = TimeIntervalParser.parse( parameterText );
             eval = new MeetsEvaluator( type,
                                        isNegated,
                                        params,
@@ -214,7 +214,7 @@ public class MeetsEvaluatorDefinition
 
         public MeetsEvaluator(final ValueType type,
                               final boolean isNegated,
-                              final Long[] parameters,
+                              final long[] parameters,
                               final String paramText) {
             super( type,
                    isNegated ? MEETS_NOT : MEETS );
@@ -330,13 +330,13 @@ public class MeetsEvaluatorDefinition
          *
          * @param parameters
          */
-        private void setParameters(Long[] parameters) {
+        private void setParameters(long[] parameters) {
             if ( parameters == null || parameters.length == 0 ) {
                 this.finalRange = 0;
             } else if ( parameters.length == 1 ) {
-                if ( parameters[0].longValue() >= 0 ) {
+                if ( parameters[0] >= 0 ) {
                     // defined max distance
-                    this.finalRange = parameters[0].longValue();
+                    this.finalRange = parameters[0];
                 } else {
                     throw new RuntimeException( "[Meets Evaluator]: Not possible to use negative parameter: '" + paramText + "'" );
                 }
